marvpred/app/requirements.txt
blinker==1.9.0
catboost==1.2.8
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
Flask==3.1.1
Flask-WTF==1.2.2
fonttools==4.59.0
graphviz==0.21
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.5
narwhals==2.0.1
numpy==1.26.4
packaging==25.0
pandas==2.3.1
pillow==11.3.0
plotly==6.2.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
rdkit-pypi==2022.9.5
scikit-learn==1.7.1
scipy==1.15.3
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.2
Werkzeug==3.1.3
WTForms==3.2.1
